+++ src/main/java/ladder/view/input/DummyView.java
@@ -1,4 +1,6 @@
-package ladder.view;
+package ladder.view.input;
+
+import ladder.view.View;
 
 public class DummyView implements View {
     @Override
